+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+ToDo:
+
+- Add AOI fields for the x and y directinos and use these in the code
+
+## 1.1.2
+
+- Refactoring the code to generate the signed 20-bit values for the waveform that are used to write it to the LC400
+
+## 1.1.1
+
+- Updated dependencies in README.md
+
 ## 1.1.0
 
 - Now uses [github/cnanders/matlab-instrument-control](https://github.com/cnanders/matlab-instrument-control) (v1.0.0-beta.*)